fix: memory.c functions use size_t instead of uint16_t.

+++ src/kernel/memory.h
@@ -1,7 +1,8 @@
 #pragma once
-#include "stdint.h"
 
-void* memcpy(void* dst, const void* src, uint16_t num);
-void* memset(void* ptr, int value, uint16_t num);
-int memcmp(const void* ptr1, const void* ptr2, uint16_t num);
+#include <stddef.h>
+
+void* memcpy(void* dst, const void* src, size_t num);
+void* memset(void* ptr, int value, size_t num);
+int memcmp(const void* ptr1, const void* ptr2, size_t num);
